refactor: extract MyWinixDeviceStub into stub.py
helpers.py imports MyWinixDeviceStub to construct device stubs, while
device_wrapper.py needs to import Helpers from helpers.py. Keeping
MyWinixDeviceStub in device_wrapper.py would create a circular import.
Moving it to a dedicated stub.py allows both modules to import freely
without a dependency cycle.
